stripe hecho otra vez de cero, ahora crea un producto en cuanto se hace la suscripción, los pagos funcionan y stripe devuelve info a la api, falta hacer algo con esa info, me voy de vacaciones

+++ src/api/models.py
@@ -68,7 +68,7 @@ class Suscription(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(24), unique=True, nullable=False)
     description = db.Column(db.String(48), unique=False, nullable=False)
-    price = db.Column(db.Float, unique=False, nullable=False)
+    price = db.Column(db.Integer, unique=False, nullable=False)
     tokens = db.Column(db.Integer, unique=False, nullable=False)
     suscription_image = db.Column(db.String(256), unique=False, nullable=True)
     suscription_type_id = db.Column(db.Integer, db.ForeignKey('suscription_type.id'), unique=False, nullable=True)
